% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_bc_name.R
\name{mascaRet_bc_name}
\alias{mascaRet_bc_name}
\title{Get the name of a boundary condition}
\usage{
mascaRet_bc_name(id, numBC)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{numBC}{: (integer, scalar) number of the boundary condition}
}
\value{
(string) the name of the boundary condition #numBC
}
\description{
Get the name of a boundary condition
}
\examples{
# mascaRet_bc_name(mascId, 2)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
