% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_bc_number.R
\name{mascaRet_bc_number}
\alias{mascaRet_bc_number}
\title{Get the number of boundary conditions}
\usage{
mascaRet_bc_number(id)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}
}
\value{
(integer, scalar) the number of boundary conditions
}
\description{
Get the number of boundary conditions
}
\examples{
# numberBC <- mascaRet_bc_number(mascId)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
