% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_computation.R
\name{mascaRet_computation}
\alias{mascaRet_computation}
\title{Computes 1D Shallow Water for a time period with new values for the boundary conditions}
\usage{
mascaRet_computation(id, verbose, tini, tend, dt, timebc, bc1, bc2)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{verbose}{: (integer, scalar) verbose mode (silent -> 0)}

\item{tini}{: (real, scalar) initial time (s)}

\item{tend}{: (real, scalar) final time (s)}

\item{dt}{: (real) time step (s)}

\item{timebc}{: (real) corresponding times of boundary condition values}

\item{bc1}{: (real) array of type 1 boundary condition values (size = length(timebc) x no. boundary conditions)}

\item{bc2}{: (real) array of type 2 boundary condition values (size = length(timebc) x no. boundary conditions)}
}
\value{
(logical, scalar) Error flag
}
\description{
Computes 1D Shallow Water for a time period with new values for the boundary conditions
}
\examples{
# mascaRet_computation(mascId, 0, 0., 1000., 10., c(0., 1000.), matrix(c(23.,23.,689,689), nrow = 2), matrix(c(23.,23.,689,689), nrow = 2))

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
