% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_compute.R
\name{mascaRet_compute}
\alias{mascaRet_compute}
\title{Computes 1D Shallow Water for a time period}
\usage{
mascaRet_compute(id, verbose, tini, tend, dt)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{verbose}{: (integer, scalar) verbose mode (silent -> 0)}

\item{tini}{: (real, scalar) initial time (s)}

\item{tend}{: (real, scalar) final time (s)}

\item{dt}{: (real, scalar) time step (s)}
}
\value{
(logical, scalar) Error flag
}
\description{
Computes 1D Shallow Water for a time period
}
\examples{
# mascaRet_compute(mascId, 0, 0., 1000., 10.)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
