% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_create.R
\name{mascaRet_create}
\alias{mascaRet_create}
\title{Creates a new instance of MASCARET}
\usage{
mascaRet_create()
}
\value{
(integer, scalar) Each call will return a new unique number (instance id)
}
\description{
Each instance is identified by unique integer number
}
\examples{
# call 'mascaRet_create'
mascaRet_problem1 <- mascaRet_create()
mascaRet_problem2 <- mascaRet_create()

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
