% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_delete.R
\name{mascaRet_delete}
\alias{mascaRet_delete}
\title{Deletes an existing instance of MASCARET}
\usage{
mascaRet_delete(id)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}
}
\value{
(logical, scalar) Error flag
}
\description{
Each instance is identified by unique integer numbers
}
\examples{
# call 'mascaRet_create'
mascaRet_problem <- mascaRet_create()
error <- mascaRet_delete(mascaRet_problem)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
