% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_error.R
\name{mascaRet_error}
\alias{mascaRet_error}
\title{Gives the error message if any}
\usage{
mascaRet_error(id)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}
}
\value{
(string) the error message
}
\description{
Gives the error message if any
}
\examples{
mascaRet_problem <- mascaRet_create()
mascaRet_error(mascaRet_problem)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
