% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_get.R
\name{mascaRet_get}
\alias{mascaRet_get}
\title{Get the value of a MASCARET variable for a particular index position}
\usage{
mascaRet_get(id, varname, i1, i2, i3)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{varname}{: (string) name of the variable to get the value from}

\item{i1}{: (integer, scalar) dimension 1 to get the value from}

\item{i2}{: (integer, scalar) dimension 2 to get the value from}

\item{i3}{: (integer, scalar) dimension 3 to get the value from}
}
\value{
(bool, integer, string or real) the value of the variable
}
\description{
Values of a variable named by 'Model.XXX' or 'State.YYY'
}
\examples{
# value <- mascaRet_get(mascId, "Model.Zbot", 24)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
