% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_hydro.R
\name{mascaRet_hydro}
\alias{mascaRet_hydro}
\title{Get the hydraulic state (Discharges et Water levels for all the river)}
\usage{
mascaRet_hydro(id)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}
}
\value{
(list) Error flag (logical) , Qi (real) , Zi (real)
}
\description{
Get the hydraulic state (Discharges et Water levels for all the river)
}
\examples{
# HydroState <- mascaRet_hydro(mascId)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
