% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_import.R
\name{mascaRet_import}
\alias{mascaRet_import}
\title{Reads a MASCARET model from files}
\usage{
mascaRet_import(id, verbose, mascfile)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{verbose}{: (integer, scalar) verbose mode (silent -> 0)}

\item{mascfile}{: (string) name of the MASCARET global file}
}
\value{
(logical, scalar) Error flag
}
\description{
All the necessary files are listed in one file ('mascfile' parameter)
}
\examples{
# mascaRet_import(mascId, 0, 'mascaret.files')

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
