% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_init.R
\name{mascaRet_init}
\alias{mascaRet_init}
\title{Initializes the MASCARET hydraulic state (Discharges 'Q' and Water levels 'Z')}
\usage{
mascaRet_init(id, q, z)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{q}{: (real) river discharges (m3/s)}

\item{z}{: (real) water levels (m)}
}
\value{
(logical, scalar) Error flag
}
\description{
(Q,Z) values for all the river (mesh nodes)
}
\examples{
# mascaRet_init(mascId, q, z)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
