% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_initialization.R
\name{mascaRet_initialization}
\alias{mascaRet_initialization}
\title{Initializes the MASCARET hydraulic state (Discharges 'Q' and Water levels 'Z') from a '.lig' file}
\usage{
mascaRet_initialization(id, filename, verbose)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{filename}{: (string) MASCARET file .lig}
}
\value{
(logical, scalar) Error flag
}
\description{
Initializes the MASCARET hydraulic state (Discharges 'Q' and Water levels 'Z') from a '.lig' file
}
\examples{
# mascaRet_init(mascId, q, z)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
