% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_set.R
\name{mascaRet_set}
\alias{mascaRet_set}
\title{Set the value of a MASCARET variable for a particular index position}
\usage{
mascaRet_set(id, varname, i1, i2, i3, value)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{varname}{: (string) name of the variable to set the value for}

\item{i1}{: (integer, scalar) dimension 1 to set the value for}

\item{i2}{: (integer, scalar) dimension 2 to set the value for}

\item{i3}{: (integer, scalar) dimension 3 to set the value for}

\item{value}{: (bool, integer, string or real) the value of the variable}
}
\value{
(logical, scalar) Error flag
}
\description{
Values of a variable named by 'Model.XXX' or 'State.YYY'
}
\examples{
# mascaRet_set(mascId, "Model.FricCoefMainCh", 12, 0, 0, 30.)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
