% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_tracer_get.R
\name{mascaRet_tracer_get}
\alias{mascaRet_tracer_get}
\title{Get values of constituents Ci}
\usage{
mascaRet_tracer_get(id)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}
}
\value{
Error flag (logical) if any or list of concentrations for each constituent C1(real) , C2(real), C3(real)...Cn(real)
}
\description{
Get values of constituents Ci
}
\examples{
# ConcTracer <- mascaRet_tracer_get(mascId)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
