% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_tracer_init.R
\name{mascaRet_tracer_init}
\alias{mascaRet_tracer_init}
\title{Initialize values of constituents}
\usage{
mascaRet_tracer_init(id, verbose)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{verbose}{: (integer, scalar) verbose mode (silent -> 0)}
}
\value{
(logical, scalar) Error flag
}
\description{
Initialize values of constituents
}
\examples{
# error <- mascaRet_tracer_init(mascId, 0)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
