% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_tracer_initialization.R
\name{mascaRet_tracer_initialization}
\alias{mascaRet_tracer_initialization}
\title{Initialize values of constituents with arrays}
\usage{
mascaRet_tracer_initialization(id, verbose, conc)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{verbose}{: (integer, scalar) verbose mode (silent -> 0)}

\item{conc}{: (real) 2D array of concentration values (number of 1D nodes x number of constituents)}
}
\value{
(logical, scalar) Error flag
}
\description{
Initialize values of constituents with arrays
}
\examples{
# error <- mascaRet_tracer_initialization(mascId, 0, C)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
