% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_varsize.R
\name{mascaRet_varsize}
\alias{mascaRet_varsize}
\title{Gives the size of a MASCARET variable}
\usage{
mascaRet_varsize(id, varname, number)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{varname}{: (string) name of the variable to get the size from}

\item{number}{: (integer, scalar) dimension 1/2/3 to get the size from}
}
\value{
(NA or integers) NA if failed or three values for each possible dimension if success
}
\description{
Sizes 1D, 2D, 3D of a variable named by 'Model.XXX' or 'State.YYY'
}
\examples{
# mascaRet_varsize(mascId, 'Model.X', 1)

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
