% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaret_vartype.R
\name{mascaRet_vartype}
\alias{mascaRet_vartype}
\title{Get the type of the MASCARET variable}
\usage{
mascaRet_vartype(id, varname)
}
\arguments{
\item{id}{: (integer, scalar) integer number identifying a MASCARET problem}

\item{varname}{: (string) name of the variable to get the type}
}
\value{
(NA or string) NA if failed or Type if success
}
\description{
Possible types are: "INT"  "DOUBLE"  "BOOL"  "STRING"  "TABINT"  "TABDOUBLE"  "TABBOOL" or "?" for unknown
}
\examples{
# Type <- mascaRet_vartype(mascId, "Model.X")

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
