% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascaRet_version.R
\name{mascaRet_version}
\alias{mascaRet_version}
\title{Get the version of TELEMAC-MASCARET}
\usage{
mascaRet_version()
}
\value{
(string) version number
}
\description{
Version is : vXpYrZ where (X,Y,Z) are numbers (see http://opentelemac.org)
}
\examples{
version <- mascaRet_version()

}
\author{
Fabrice Zaoui - Copyright EDF 2020
}
